# dnsproxy: split MatchPattern and MatchName to improve perf