--- a/pkg/fqdn/dnsproxy/helpers_test.go
+++ b/pkg/fqdn/dnsproxy/helpers_test.go
@@ -6,6 +6,7 @@
 package dnsproxy
 
 import (
+	"regexp"
 	"testing"
 
 	. "gopkg.in/check.v1"
@@ -30,7 +31,8 @@ func (s *DNSProxyHelperTestSuite) TestSetPortRulesForID(c *C) {
 	re.InitRegexCompileLRU(1)
 	rules := policy.L7DataMap{}
 	epID := uint64(1)
-	pea := newPerEPAllow()
+	pea := perEPAllow{}
+	cache := newMatcherCache()
 	rules[new(MockCachedSelector)] = &policy.PerSelectorPolicy{
 		L7Rules: api.L7Rules{
 			DNS: []api.PortRuleDNS{
@@ -39,10 +41,10 @@ func (s *DNSProxyHelperTestSuite) TestSetPortRulesForID(c *C) {
 			},
 		},
 	}
-	err := pea.setPortRulesForID(epID, 8053, rules)
+	err := pea.setPortRulesForID(cache, epID, 8053, rules)
 	c.Assert(err, Equals, nil)
-	c.Assert(len(pea.patternMatchersByPattern), Equals, 1)
-	c.Assert(len(pea.nameMatcherBySignature), Equals, 1)
+	c.Assert(len(cache.patternMatcherCache), Equals, 1)
+	c.Assert(len(cache.nameMatcherCache), Equals, 1)
 
 	selector2 := new(MockCachedSelector)
 	rules[selector2] = &policy.PerSelectorPolicy{
@@ -54,25 +56,68 @@ func (s *DNSProxyHelperTestSuite) TestSetPortRulesForID(c *C) {
 			},
 		},
 	}
-	err = pea.setPortRulesForID(epID, 8053, rules)
+	err = pea.setPortRulesForID(cache, epID, 8053, rules)
 	c.Assert(err, Equals, nil)
-	c.Assert(len(pea.patternMatchersByPattern), Equals, 2)
-	c.Assert(len(pea.nameMatcherBySignature), Equals, 2)
+	c.Assert(len(cache.patternMatcherCache), Equals, 2)
+	c.Assert(len(cache.nameMatcherCache), Equals, 2)
 
 	delete(rules, selector2)
-	err = pea.setPortRulesForID(epID, 8053, rules)
+	err = pea.setPortRulesForID(cache, epID, 8053, rules)
 	c.Assert(err, Equals, nil)
-	c.Assert(len(pea.patternMatchersByPattern), Equals, 1)
-	c.Assert(len(pea.nameMatcherBySignature), Equals, 1)
+	c.Assert(len(cache.patternMatcherCache), Equals, 1)
+	c.Assert(len(cache.nameMatcherCache), Equals, 1)
 
-	err = pea.setPortRulesForID(epID, 8053, nil)
+	err = pea.setPortRulesForID(cache, epID, 8053, nil)
 	c.Assert(err, Equals, nil)
-	c.Assert(len(pea.patternMatchersByPattern), Equals, 0)
-	c.Assert(len(pea.nameMatcherBySignature), Equals, 0)
-
+	c.Assert(len(cache.patternMatcherCache), Equals, 0)
+	c.Assert(len(cache.nameMatcherCache), Equals, 0)
 }
 
-func (s *DNSProxyHelperTestSuite) TestGenerateRegexpAndFqdns(c *C) {
+func (s *DNSProxyHelperTestSuite) TestSetPortRulesForIDFromUnifiedFormat(c *C) {
+	re.InitRegexCompileLRU(1)
+	rules := make(CachedSelectorREEntry)
+	epID := uint64(1)
+	pea := perEPAllow{}
+	cache := newMatcherCache()
+	rules[new(MockCachedSelector)] = regexp.MustCompile("^.*[.]cilium[.]io$")
+	rules[new(MockCachedSelector)] = regexp.MustCompile("^.*[.]cilium[.]io$")
+
+	err := pea.setPortRulesForIDFromUnifiedFormat(cache, epID, 8053, rules)
+	c.Assert(err, Equals, nil)
+	c.Assert(len(cache.patternMatcherCache), Equals, 1)
+	c.Assert(len(cache.nameMatcherCache), Equals, 0)
+
+	selector2 := new(MockCachedSelector)
+	rules[selector2] = regexp.MustCompile("^sub[.]cilium[.]io")
+	err = pea.setPortRulesForIDFromUnifiedFormat(cache, epID, 8053, rules)
+	c.Assert(err, Equals, nil)
+	c.Assert(len(cache.patternMatcherCache), Equals, 2)
+	c.Assert(len(cache.nameMatcherCache), Equals, 0)
+
+	delete(rules, selector2)
+	err = pea.setPortRulesForIDFromUnifiedFormat(cache, epID, 8053, rules)
+	c.Assert(err, Equals, nil)
+	c.Assert(len(cache.patternMatcherCache), Equals, 1)
+	c.Assert(len(cache.nameMatcherCache), Equals, 0)
+
+	err = pea.setPortRulesForIDFromUnifiedFormat(cache, epID, 8053, nil)
+	c.Assert(err, Equals, nil)
+	c.Assert(len(cache.patternMatcherCache), Equals, 0)
+	c.Assert(len(cache.nameMatcherCache), Equals, 0)
+
+	delete(rules, selector2)
+	err = pea.setPortRulesForIDFromUnifiedFormat(cache, epID, 8053, rules)
+	c.Assert(err, Equals, nil)
+	c.Assert(len(cache.patternMatcherCache), Equals, 1)
+	c.Assert(len(cache.nameMatcherCache), Equals, 0)
+
+	err = pea.setPortRulesForIDFromUnifiedFormat(cache, epID, 8053, nil)
+	c.Assert(err, Equals, nil)
+	c.Assert(len(cache.patternMatcherCache), Equals, 0)
+	c.Assert(len(cache.nameMatcherCache), Equals, 0)
+}
+
+func (s *DNSProxyHelperTestSuite) TestGeneratePatternAndFqdns(c *C) {
 	dnsName := "example.name."
 	dnsPattern := "*matc*.name."
 
@@ -87,9 +132,9 @@ func (s *DNSProxyHelperTestSuite) TestGenerateRegexpAndFqdns(c *C) {
 		}},
 	}
 	re.InitRegexCompileLRU(defaults.FQDNRegexCompileLRUSize)
-	regex, fqdns := GenerateRegexpAndFqdns(l7)
+	pattern, fqdns := GeneratePatternAndFQDNs(l7)
 
-	regexp, err := re.CompileRegex(regex)
+	regex, err := re.CompileRegex(pattern)
 	c.Assert(err, Equals, nil)
 
 	domainMatcher := func(fqdn string) bool {
@@ -98,7 +143,7 @@ func (s *DNSProxyHelperTestSuite) TestGenerateRegexpAndFqdns(c *C) {
 				return true
 			}
 		}
-		return regexp.MatchString(fqdn)
+		return regex.MatchString(fqdn)
 	}
 
 	c.Assert(domainMatcher(dnsName), Equals, true)