@odinuge 2022-07-15
I've also updated my benchmark code btw., and made a proper README highlighting how the various things improve the perf; https://github.com/odinuge/cilium-fqdn-regex-benchmarking/blob/master/README.md