@odinuge 2022-08-05
Thanks for all the reviews!

Some updates;
- Updated a lot of the godocs as suggested by @aanm and @christarazi
- Fixed the `UpdateAllowedFromSelectorRegexes` func
- Added support for `nil` regex that means it matches no fqdns, same for the name matcher part as suggested by @christarazi and @jrajahalme
- Extracted the "cache" out to a separate variable to make it easier to mange and test
- Added the restored rules regex to the cache to avoid having to recompile it again when the "normal" rules are inserted
  - We can probably do the same for the matchName map as well, but with the `map[string]struct{}` format on disk its a bit awkward due to the sorting. We can however do that in another flup PR; and the benefit for the regex is much greater!
  - We can also avoid compiling the regex in `pkg/fqdn/restore/restore.go` and just pass the nullable string to the dnsproxy. That would improve startup perf when the same regex is used multiple times.
- Added a few small tests as well, but maybe we need more.
- Pushed a `--fixup` commit to make it easer to read the delta. Will squash before merging when that time comes.

Ref. the comment about backwards compatibility, now that the storage format for the `Re` value can be `null`, rollback _will_ break, so we need to backport a fix to support that, and at the same time support the new `FQDNs`.

Say that we do this (as an example);
- We ship this PR in v1.13.0
- We backport a fix to v1.12.1

Then, if people upgrade from v1.12.0 to v1.13.0 and roll back, it will break, and they will end up with a null pointer exception, right?. That is not ideal if thats the case.
- Maybe we can use empty string on the storage format as a temp value in v1.13.0, and move to an actual null in v1.14.0? That way people don't break when going from v1.13.0->v1.12.0, even though they will "loose" the restore part for the new FQDNs map.
- Or do we have to wait for v1.14.0 to land this PR?

Will take another look tomorrow/early next week and go through all the phrasing as suggested by @christarazi.

Thanks
Odin