@christarazi 2022-08-05
@odinuge I think the ideal approach would be to file a separate PR to land the backwards compat changes so that we can backport it. Then in this PR it builds on top of it. I don't think we need to wait until 1.14. We should be able to backport the first PR into v1.12.x and land this one in 1.13.