@odinuge 2022-08-12
Chatted with @christarazi offline about this. Due to complexity getting this backported, we decided to go for ~https://github.com/cilium/cilium/compare/master...odinuge:ou-testing-only-regex for now, and do this approach after we have gotten that initial change merged and back ported. I'm out for two weeks, but will attack it when I'm back. Thanks, Odin.